requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==4.2.6
django-admin-confirm==1.0.0
django-admin-extra-buttons==1.5.6
django-admin-shortcuts==2.1.1
django-filter==23.3
django-import-export==3.3.1
django-jazzmin==2.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
et-xmlfile==1.1.0
idna==3.4
Markdown==3.5
MarkupPy==1.14
mssql-django==1.3
numpy==1.26.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.1
Pillow==10.0.1
pycparser==2.21
PyJWT==2.8.0
pymssql==2.2.8
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
urllib3==2.0.6
xlrd==2.0.1
xlwt==1.3.0
